requirements.txt
attrs>=21,<24
requests
cookiecutter~=2.6.0
click~=8.1.0
pyyaml
terminaltables
cloudshell-rest-api~=9.0.0
colorama
giturlparse
ruamel.yaml
cryptography
setuptools; python_version >= '3.12'
packaging